+++ pcs/lib/cib/fencing_topology.py
@@ -1,4 +1,5 @@
 from typing import (
+    Iterable,
     Optional,
     Sequence,
     Set,
@@ -97,20 +98,24 @@ def remove_all_levels(topology_el):
 def remove_levels_by_params(
     topology_el: _Element,
     level=None,
-    target_type=None,
+    # TODO create a special type, so that it cannot accept any string
+    target_type: Optional[str] = None,
     target_value=None,
-    devices=None,
-    ignore_if_missing=False,
+    devices: Optional[Iterable[str]] = None,
+    # TODO remove, deprecated backward compatibility layer
+    ignore_if_missing: bool = False,
+    # TODO remove, deprecated backward compatibility layer
+    validate_device_ids: bool = True,
 ) -> ReportItemList:
     """
     Remove specified fencing level(s)
 
-    etree topology_el -- etree element to remove the levels from
+    topology_el -- etree element to remove the levels from
     int|string level -- level (index) of the fencing level to remove
-    constant target_type -- the removed fencing level target value type
+    target_type -- the removed fencing level target value type
     mixed target_value -- the removed fencing level target value
-    Iterable devices -- list of stonith devices of the removed fencing level
-    bool ignore_if_missing -- when True, do not report if level not found
+    devices -- list of stonith devices of the removed fencing level
+    ignore_if_missing -- when True, do not report if level not found
     """
     # Do not ever remove a fencing-topology element, even if it is empty. There
     # may be ACLs set in pacemaker which allow "write" for fencing-level
@@ -125,6 +130,12 @@ def remove_levels_by_params(
         if has_errors(report_list):
             return report_list
 
+    if validate_device_ids and devices is not None:
+        for device_id in devices:
+            validate_id(
+                device_id, description="stonith id", reporter=report_list
+            )
+
     level_el_list = _find_level_elements(
         topology_el, level, target_type, target_value, devices
     )
@@ -135,7 +146,10 @@ def remove_levels_by_params(
         report_list.append(
             ReportItem.error(
                 reports.messages.CibFencingLevelDoesNotExist(
-                    level, target_type, target_value, devices or []
+                    level,
+                    target_type,
+                    target_value,
+                    sorted(devices) if devices else [],
                 )
             )
         )